requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
backoff==2.2.1
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.13
databricks-sdk==0.67.0
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162070/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
docker==7.1.0
ensure==1.0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
-e git+https://github.com/nhut-nam/FaceEmotionRecognitionSystemTest.git@a549bf5b2e0e28c11e6e1ca106ab979d360fad97#egg=FacialExpressionRecognition
fastapi==0.117.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.0
fsspec==2025.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mlflow==2.22.2
mlflow-skinny==2.22.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy==2.2.6
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==24.2
pandas==2.3.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pathvalidate==3.3.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.3.2
protobuf==6.32.1
psutil @ file:///D:/bld/psutil_1758169180033/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
PySocks==1.7.1
python-box==7.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
PyYAML==6.0.3
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.8.0
torchvision==0.23.0
tornado @ file:///D:/bld/tornado_1756854837538/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
waitress==3.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
Werkzeug==3.1.3
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
